Add files via upload
This file contain all 3D CAD drawing of the eight stepped hulls. These hulls are in scale 1:1 respect the models of resistance tests with a displacement at the time of tests of 3,130 Kg in water with a density of 1000 Kg/m^3. Tests are developed in Towing Tank of University of Napoli "Federico II" Dept. of Industrial Engineering.